requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
defusedxml==0.7.1
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-xml==2.0.0
djangorestframework_simplejwt==5.5.1
idna==3.11
oauthlib==3.3.1
pillow==12.1.0
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.5
tweepy==4.16.0
urllib3==2.6.3
